issue service update

+++ src/main/java/com/sourcery/oirs/database/entity/IssueEntity.java
@@ -1,15 +1,10 @@
 package com.sourcery.oirs.database.entity;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.sourcery.oirs.model.Issue;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
-
 import java.sql.Timestamp;
-import java.time.LocalDate;
 import java.util.UUID;
 
 @NoArgsConstructor
@@ -22,11 +17,11 @@ public class IssueEntity {
     private String name;
     private String status;
     private String description;
-    private Long commentCount;
+    private Double commentCount;
     private Timestamp startTime;
     private Timestamp finishTime;
     private UUID employeeId;
     private UUID officeId;
-    private Long rating;
+    private Double rating;
 
-}
+    }
